add completed general section, outline content components
This commit creates the general section and adds functionality/content
to the general section. The user can press the pencil button to swap
the text field to a form input field. When the input field is modified
the state is automatically updated with the new information. The other
four components skills, education, work experience and certifications
have been briefly outlined.