fixed bug where skills section props were not updating
It turns out I was implimenting the skills container incorrectly.
When the state was updated and passed to the component react was
attempting to set the state without using this.setState() (in the class
component constructor). The correct implimentation was to create a
function for the component and export it as default. I then was able
to access the updated props by using the props argument.