add dynamically generated stars, add remove image
This commit adds a function getStars which dynamically generate
jsx for the amount of stars the user has inputted for a skill.
The function also prettifies the display by setting the half star
in between any filled stars and star outlines (if applicable).
A remove image was added next to each skill however an onClick function
still needs to be created to remove the skill.