add remove skill functionality
This commit adds a method removeSkill passed as a prop to the skill
container. .bind(this) was added in the prop passed to avoid
any state is not defined errors when calling the method in a different
context. A skill can be removed by pressing the - button next to the
skill in the list.