add toggle function for certs, add remove cert function
This commit adds a function which toggles the form to update a cert
by pressing the pencil (edit) button next to the cert. Functionality
has not been added to update the state of the cert with the updated
field data. A method to remove certs from the certArray was also created
and is passed to the certcontainer as a prop. In order to allow
the state to be updated in the child component when a cert is removed
the cert is passed with .bind(this).