Impliment Education & EducationContainer components
This commit creates two components Education and Education Container.
The Education Container component houses all of the schools the
user has attended. By pressing the edit button the user will be prompted
with input fields to modify the selected education item. The inputs
will automatically update the state of the selected education item.