Update new item placeholders, generally position components
This commit changes the names of the new items in the education, skills,
work experience and certifications sections to descriptions of what
that field should contain. This commit also adjusts the location
of components to roughly where they should be in addition to adding
light styling to the font sizes.