add css positioning using grid and flex
This commit adds the finishing touches to forms when the edit button is
pressed on a resume item. Labels and their fields are also now
properly organized. Prior to this commit labels would appear in
incorrect positions. The positioning of some form field toggles
had to be changed to grid instead of block to ensure proper positioning.